# docs/node_swagger.yaml
definitions:
...
  data.AccountInfo:
    properties:
      address:
        type: string
      allowance:
        type: integer
      balance:
        type: integer
      codeHash:
        type: string
      codeMetadata:
        type: string
      foundation:
        type: boolean
      frozenBalance:
        type: integer
      name:
        type: string
      nonce:
        type: integer
      permissions:
        items:
          $ref: '#/definitions/data.Permissions'
        type: array
      rootHash:
        type: string
      timestamp:
        $ref: '#/definitions/time.Duration'
      unfrozenBalance:
        type: integer
    type: object
...
  data.PermissionKey:
    properties:
      address:
        type: string
      weight:
        type: integer
    type: object
  data.Permissions:
    properties:
      Threshold:
        type: integer
      id:
        type: integer
      operations:
        type: string
      permissionName:
        type: string
      signers:
        items:
          $ref: '#/definitions/data.PermissionKey'
        type: array
      type:
        type: integer
    type: object
...
  kapps.LastClaim:
    properties:
      Epoch:
        type: integer
      Timestamp:
        type: integer
    type: object
...
  kapps.UserBucket:
    properties:
      Delegation:
        items:
          type: integer
        type: array
      StakedAt:
        type: integer
      StakedEpoch:
        type: integer
      UnstakedEpoch:
        type: integer
      Value:
        type: integer
    type: object
  kapps.UserKDA:
    properties:
      Balance:
        type: integer
      Buckets:
        additionalProperties:
          $ref: '#/definitions/kapps.UserBucket'
        type: object
      FrozenBalance:
        type: integer
      LastClaim:
        $ref: '#/definitions/kapps.LastClaim'
      MIME:
        items:
          type: integer
        type: array
      Metadata:
        items:
          type: integer
        type: array
    type: object
...
  models.AccountNonceResponse:
    properties:
      firstPendingNonce:
        type: integer
      nonce:
        type: integer
      txPending:
        type: integer
    type: object
  models.AvailableClaimListResponse:
    properties:
      assets:
        additionalProperties:
          $ref: '#/definitions/models.AvailableClaimResponse'
        type: object
    type: object
  models.AvailableClaimResponse:
    properties:
      allStakingRewards:
        additionalProperties:
          format: int64
          type: integer
        type: object
      allowance:
        type: integer
      stakingRewards:
        type: integer
    type: object
  models.BalanceResponse:
    properties:
      balance:
        type: integer
    type: object
...
  models.KDAResponse:
    properties:
      address:
        type: string
      asset:
        type: string
      userKDA:
        $ref: '#/definitions/kapps.UserKDA'
    type: object
...
  shared.GenericAPIResponse:
    properties:
      code:
        $ref: '#/definitions/shared.ReturnCode'
      data: {}
      error:
        type: string
    type: object
  shared.ReturnCode:
    enum:
    - successful
    - internal_issue
    - bad_request
    - system_busy
    type: string
    x-enum-varnames:
    - ReturnCodeSuccess
    - ReturnCodeInternalError
    - ReturnCodeRequestError
    - ReturnCodeSystemBusy
...
  time.Duration:
    enum:
    - -9223372036854775808
    - 9223372036854775807
    - 1
    - 1000
    - 1000000
    - 1000000000
    - 60000000000
    - 3600000000000
    format: int64
    type: integer
    x-enum-varnames:
    - minDuration
    - maxDuration
    - Nanosecond
    - Microsecond
    - Millisecond
    - Second
    - Minute
    - Hour
...
info:
  contact: {}
  description: This is a sample of the KleverChain Node Server.
  title: KleverChain Node
  version: "1.0"
paths:
  /address/{address}:
    get:
      parameters:
      - description: address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  properties:
                    account:
                      $ref: '#/definitions/data.AccountInfo'
                  type: object
              type: object
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: returns an accountResponse
      tags:
      - Address
  /address/{address}/allowance:
    get:
      parameters:
      - description: address
        in: path
        name: address
        required: true
        type: string
      - description: asset
        in: query
        name: asset
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.AvailableClaimResponse'
              type: object
        "400":
          description: some error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: returns the rewards available for a specific asset in an account
      tags:
      - Address
  /address/{address}/allowance/list:
    get:
      parameters:
      - description: address
        in: path
        name: address
        required: true
        type: string
      - description: assets (comma-separated asset IDs, max 25), e.g. KLV,KFI
        in: query
        name: asset
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.AvailableClaimListResponse'
              type: object
        "400":
          description: some error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: returns the rewards available for a specific list of asset in an account
      tags:
      - Address
  /address/{address}/balance:
    get:
      parameters:
      - description: address
        in: path
        name: address
        required: true
        type: string
      - default: KLV
        description: asset
        in: query
        name: asset
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.BalanceResponse'
              type: object
        "400":
          description: some error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: returns the balance for the address parameter
      tags:
      - Address
  /address/{address}/kda:
    get:
      parameters:
      - description: address
        in: path
        name: address
        required: true
        type: string
      - default: KLV
        description: asset
        in: query
        name: asset
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.KDAResponse'
              type: object
        "400":
          description: some error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: returns the user kda for the address parameter
      tags:
      - Address
  /address/{address}/nonce:
    get:
      parameters:
      - description: address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/shared.GenericAPIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.AccountNonceResponse'
              type: object
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/shared.GenericAPIResponse'
      summary: returns the nonce and pending transaction info for a specific account
      tags:
      - Address
...
swagger: "2.0"